Created draft of OfferDTO

+++ src/main/java/com/example/dorywcza/model/offer/Offer.java
@@ -1,6 +1,5 @@
 package com.example.dorywcza.model.offer;
 
-import com.example.dorywcza.model.Industry;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.sun.istack.NotNull;
 import lombok.AllArgsConstructor;
@@ -21,7 +20,7 @@ public class Offer {
     private Long id;
     private String title;
     private String description;
-    private int userId;
+    private Long userId;
     private boolean hasExperience;
 
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd", timezone = "CET")
@@ -64,8 +63,8 @@ public class Offer {
     private OfferSchedule offerSchedule;
 
 
-    public Offer(String title, String description, int userId, boolean hasExperience, OfferLocation offerLocation,
-                 DateRange dateRange, Industry industry,  Salary salary, OfferSchedule offerSchedule) {
+    public Offer(String title, String description, Long userId, boolean hasExperience, OfferLocation offerLocation,
+                 DateRange dateRange, Industry industry, Salary salary, OfferSchedule offerSchedule) {
         this.title = title;
         this.description = description;
         this.userId = userId;
@@ -75,6 +74,6 @@ public class Offer {
         this.dateRange = dateRange;
         this.industry = industry;
         this.offerSchedule = offerSchedule;
-
     }
+
 }
