+++ src/main/java/com/example/dorywcza/model/service_offer/ServiceOffer.java
@@ -1,7 +1,7 @@
 package com.example.dorywcza.model.service_offer;
 
 
-import com.example.dorywcza.model.Industry;
+import com.example.dorywcza.model.offer.Industry;
 import com.example.dorywcza.model.offer.*;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -16,7 +16,7 @@ public class ServiceOffer extends Offer {
 
 
 
-    public ServiceOffer(String title, String description, int userId, boolean hasExperience,
+    public ServiceOffer(String title, String description, Long userId, boolean hasExperience,
                         OfferLocation offerLocation, DateRange dateRange, Industry industry,
                         Salary salary, OfferSchedule offerSchedule) {
         super(title, description, userId, hasExperience, offerLocation, dateRange, industry,  salary, offerSchedule);
